% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_is_spline.R
\name{is.splinets,Splinets-method}
\alias{is.splinets,Splinets-method}
\title{Diagnostics of splines}
\usage{
\S4method{is.splinets}{Splinets}(object)
}
\arguments{
\item{object}{\code{Splinets} object, the object to be diagnosed;}
}
\description{
This short information is added to satisfy an R-package building requirement, see \code{\link{is.splinets}} 
for the actual information.
}
